+++ react/components/SKUSelector/utils/index.js
@@ -2,13 +2,17 @@ import { clone } from 'ramda'
 
 /**
  * Return the maximum sku price
- * @param {array of sku's} items 
+ * @param {array of sku's} items
  */
 export const getMaxSkuPrice = items => {
   if (!items) return 0
 
   return items.reduce((max, sku) => {
-    const [{ commertialOffer: { Price } }] = sku.sellers
+    const [
+      {
+        commertialOffer: { Price },
+      },
+    ] = sku.sellers
 
     return Math.max(max, Price)
   })
@@ -16,18 +20,18 @@ export const getMaxSkuPrice = items => {
 
 /**
  * Remove the 'https' from the given url
- * @param {string} url 
+ * @param {string} url
  */
 export const stripUrl = url => url.replace(/^https?:/, '')
 
 /**
  * Parse the variations field in the sku object
- * @param {sku} sku 
+ * @param {sku} sku
  */
-export const parseSku = (sku) => {
+export const parseSku = sku => {
   const result = clone(sku)
 
-  const variations = sku.variations.map((variation) => {
+  const variations = sku.variations.map(variation => {
     result[variation.name] = variation.values[0]
     return variation.name
   })
@@ -39,13 +43,13 @@ export const parseSku = (sku) => {
 
 /**
  * Retrieves a list of unique options of a given variation
- * @param {string} variation 
- * @param {skus} skus 
+ * @param {string} variation
+ * @param {skus} skus
  */
 export const getVariationOptions = (variation, skus) => {
   const hTable = {}
 
-  skus.map((sku) => {
+  skus.map(sku => {
     const value = sku[variation]
     hTable[value] = sku
   })
@@ -59,19 +63,21 @@ export const getVariationOptions = (variation, skus) => {
 
 /**
  * Verifies if the variation is color
- * @param {string} variation 
+ * @param {string} variation
  */
-export const isColor = (variation) => {
+export const isColor = variation => {
   if (!variation) return false
 
-  return variation.toLowerCase() === "cor" || variation.toLowerCase() === "color"
+  return (
+    variation.toLowerCase() === 'cor' || variation.toLowerCase() === 'color'
+  )
 }
 
 /**
  * Choose wich variation will be the main one.
- * @param {Array[string]} variations 
+ * @param {Array[string]} variations
  */
-export const getMainVariationName = (variations) => {
+export const getMainVariationName = variations => {
   for (let i = 0; i < variations.length; i++) {
     if (isColor(variations[i])) return variations[i]
   }