+++ react/components/UserAddress/index.js
@@ -3,8 +3,9 @@ import React from 'react'
 import { FormattedMessage } from 'react-intl'
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
+import { compose, pathOr } from 'ramda'
 
-import ChangeAddressIcon from './ChangeAddressIcon'
+import { IconLocationMarker } from 'vtex.store-icons'
 import Container from '../Container'
 
 class UserAddress extends React.Component {
@@ -31,9 +32,12 @@ class UserAddress extends React.Component {
         ? `${displayStreet} - ${complement}`
         : `${displayStreet}`
 
-    /** TODO: get pickupPoint name instead of address, if it's pickup
-     * @author lbebber */
     const isPickup = addressType === 'pickup'
+    const friendlyName = pathOr(
+      '',
+      ['pickupPointQuery', 'pickupPoint', 'friendlyName'],
+      this.props
+    )
 
     return (
       <div
@@ -45,7 +49,7 @@ class UserAddress extends React.Component {
               inverted ? 'c-on-base--inverted' : 'c-muted-2'
             }`}
           >
-            <ChangeAddressIcon />
+            <IconLocationMarker size={27} viewBox={'0 0 21 27'} />
           </div>
           <div className="flex flex-auto flex-column">
             <div
@@ -54,7 +58,10 @@ class UserAddress extends React.Component {
               }`}
             >
               {isPickup ? (
-                <FormattedMessage id="user-address.pickup" />
+                <FormattedMessage
+                  id="user-address.pickup"
+                  values={{ name: friendlyName }}
+                />
               ) : (
                 <FormattedMessage id="user-address.order" />
               )}
@@ -125,44 +132,43 @@ class UserAddress extends React.Component {
   }
 }
 
-export default graphql(
+const withShippingDataQuery = graphql(
   gql`
     query {
       minicart @client {
-        orderForm {
-          shippingData {
-            address {
-              id
-              neighborhood
-              complement
-              number
-              street
-              postalCode
-              city
-              reference
-              addressName
-              addressType
-            }
-            availableAddresses {
-              id
-              neighborhood
-              complement
-              number
-              street
-              postalCode
-              city
-              reference
-              addressName
-              addressType
-            }
-          }
-        }
+        orderForm
       }
     }
   `,
   {
     props: ({ data: { minicart } }) => ({
-      orderForm: minicart && minicart.orderForm || {},
+      orderForm:
+        minicart && minicart.orderForm ? JSON.parse(minicart.orderForm) : {},
     }),
   }
+)
+
+const withPickupPointQuery = graphql(
+  gql`
+    query pickupPoint($id: String!) {
+      pickupPoint(id: $id) {
+        friendlyName
+      }
+    }
+  `,
+  {
+    skip: ({ orderForm: { checkedInPickupPointId, isCheckedIn } }) =>
+      !checkedInPickupPointId || !isCheckedIn,
+    options: ({ orderForm: { checkedInPickupPointId } }) => ({
+      variables: {
+        id: checkedInPickupPointId,
+      },
+    }),
+    name: 'pickupPointQuery',
+  }
+)
+
+export default compose(
+  withShippingDataQuery,
+  withPickupPointQuery
 )(UserAddress)